+++ pbparam/optimisation_problems/data_fit.py
@@ -1,7 +1,8 @@
 #
-# Optimisation problem class
+# Data fit class
 #
 
+import pbparam
 import pybamm
 import numpy as np
 import copy
@@ -48,7 +49,7 @@ def cost_function_full(simulation, map_inputs, data, x):
     return np.array(NRMSE)
 
 
-class OptimisationProblem(object):
+class DataFit(pbparam.BaseOptimisationProblem):
     """A class to define an optimisation problem.
 
     Parameters
@@ -65,9 +66,10 @@ class OptimisationProblem(object):
         will take the same value.
         TODO: allow it to be a list of variables and pass the bounds and initial
         guesses when running the optimiser.
-    variables_to_optimise : str or list of str (optional)
+    variables_optimise : str or list of str (optional)
         The variable or variables to optimise in the cost function. The default is
         "Terminal voltage [V]". It can be a string or a list of strings.
+        TODO: implement this
     """
 
     def __init__(
@@ -137,3 +139,63 @@ class OptimisationProblem(object):
             cost_function_full, simulation, self.map_inputs, self.data
         )
         self.cost_function = cost_function
+
+    def calculate_solution(self, parameters=None):
+        """
+        Calculate solution of model.
+
+        Parameters
+        ----------
+        parameters : array-like (optional)
+            Parameters to calcualte the solution for. If not provided, it uses the
+            original parameters.
+
+        Returns
+        -------
+        solution : pybamm.Solution
+            The solution for the given inputs.
+        """
+        if parameters:
+            inputs = {param: parameters[i] for param, i in self.map_inputs.items()}
+        else:
+            inputs = {
+                param: self.original_parameters[param]
+                for param in self.map_inputs.keys()
+            }
+
+        if getattr(self.simulation, "experiment", None):
+            t_eval = None
+        else:
+            t_eval = [0, self.data["Time [s]"].iloc[-1]]
+
+        solution = self.simulation.solve(t_eval=t_eval, inputs=inputs)
+
+        return solution
+
+    def _plot(self, x_optimal):
+        """
+        Plot the optimisation result. Should be accessed through the OptimisationResult
+        plot method.
+        """
+
+        initial_solution = self.calculate_solution()
+        optimal_solution = self.calculate_solution(x_optimal)
+
+        plot = pybamm.QuickPlot(
+            [initial_solution, optimal_solution],
+            output_variables=self.variables_optimise,
+            labels=["Initial values", "Optimal values"],
+        )
+
+        plot.plot(0)
+
+        for ax, var in zip(plot.axes, self.variables_optimise):
+            data = self.data
+            ax.plot(
+                data["Time [s]"] / plot.time_scaling_factor,
+                data[var],
+                "k:",
+                label="Data"
+            )
+
+        return plot